requirements.txt
mysql-connector-python
openai-whisper
python-crontab
schedule
ffmpeg-python==0.2.0